+++ package.json
@@ -28,6 +28,7 @@
     "@trpc/next": "^10.21.1",
     "@trpc/react-query": "^10.21.1",
     "@trpc/server": "^10.9.0",
+    "@types/lodash": "^4.14.194",
     "@uiball/loaders": "^1.2.6",
     "@upstash/ratelimit": "^0.4.2",
     "@vercel/analytics": "^1.0.0",
@@ -41,6 +42,7 @@
     "html-to-image": "^1.11.11",
     "i18next": "^22.4.15",
     "langchain": "^0.0.63",
+    "lodash": "^4.17.21",
     "micro": "^10.0.1",
     "micro-cors": "^0.1.1",
     "next": "13.1.6",
