+++ src/pages/api/auth/[...nextauth].ts
@@ -1,9 +1,10 @@
 import NextAuth from "next-auth";
-import { options } from "../../../server/auth";
+import { authOptions } from "../../../server/auth";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-const auth = (req: NextApiRequest, res: NextApiResponse) =>
+const auth = (req: NextApiRequest, res: NextApiResponse) => {
   // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-  NextAuth(req, res, options(req, res));
+  return NextAuth(req, res, authOptions(req, res));
+};
 
 export default auth;
